chatbot/surgical-cbr/requirements.txt
fastapi
uvicorn[standard]
pydantic
qdrant-client[qdrant]   # <-- embedded version
google-generativeai
python-dotenv
requests

